Se implementa servicio deleteById a productos con graphql

+++ Desafio19/controller/productsController.js
@@ -82,6 +82,12 @@ export async function updateProductById(req, res) {
 
 export async function doDeleteProductById(req, res) {
     const {id} = req.params;
-    let deletedProduct = await service.deleteProductById(id);
-    res.send(deletedProduct);
+    let graphqlQuery = req.originalUrl.search('graphql') > 0 || false;
+    if(graphqlQuery){ //GraphQl resuelve promesas.
+        let deletedProduct = await service.deleteProductById(id);
+        console.log(deletedProduct);
+        return deletedProduct;
+    }else{
+        res.send(deletedProduct);
+    }
 }
